requirements.txt
paramiko~=3.5.0
schedule~=1.2.2
requests~=2.32.3
python-dotenv~=1.0.1
loguru~=0.7.3